+++ app/src/main/java/ly/count/android/demo/App.java
@@ -181,10 +181,10 @@ public class App extends Application {
             .enableAutomaticViewTracking()
             // uncomment the line below to enable auto enrolling the user to AB experiments when downloading RC data
             //.enrollABOnRCDownload()
+            // .setMaxRequestQueueSize(5)
             .setAutoTrackingUseShortName(true)
             .setAutomaticViewSegmentation(automaticViewSegmentation)
             .setAutoTrackingExceptions(new Class[] { ActivityExampleCustomEvents.class })
-
 
             .setPushIntentAddMetadata(true)
 
@@ -197,7 +197,7 @@ public class App extends Application {
             //.enableTemporaryDeviceIdMode()
 
             .setRequiresConsent(true)
-            
+
             //for giving all consent values
             .giveAllConsents()
 
@@ -210,6 +210,7 @@ public class App extends Application {
             //})
 
             .setHttpPostForced(false)
+            // .setRequestDropAgeHours(10) // set this if you want to drop any request that was not sent in a certain timeframe
             .setParameterTamperingProtectionSalt("test-salt-checksum")
             .addCustomNetworkRequestHeaders(customHeaderValues)
             //.enableCertificatePinning(certificates)