+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## X.X.X
+* Added a config method to set a time limit after which the requests would be removed if not sent to the server: 'setRequestDropAgeHours'
+
 ## 23.8.2
 * Fixed APM bug where automatic foreground, background tracking would track wrong if the SDK was not initialized while the app was not in the foreground